+++ jax/_src/pallas/mosaic_gpu/primitives.py
@@ -3737,11 +3737,28 @@ def _async_store_tmem_lowering_rule(
   x_tmem, _, transforms = lowering._handle_transforms(
       ctx, x_aval, x_ref, transform_avals, transforms, handle_transposes=False,
       handle_reshapes=False)
+  batch_shape = ()
+  if transforms and isinstance(
+      transforms[0], gpu_core.ExpandLeadingBatchDimensionsTransform
+  ):
+    batch_shape = transforms[0].batch_shape
+    transforms = transforms[1:]
   if transforms:
     raise NotImplementedError(
         f"Unimplemented transforms for TMEM refs. {transforms=}"
     )
-  x_tmem.store(value)
+  if batch_shape:
+    m, n = value.shape[-2:]
+    for batch_idx in np.ndindex(batch_shape):
+      flat_batch_idx = int(np.ravel_multi_index(batch_idx, batch_shape))
+      # TODO(allanrenucci): Add direct support for indexing to FragmentedArray.
+      slices = tuple(slice(i, i + 1) for i in batch_idx)
+      val_slice = value[slices].reshape((m, n))
+      col_start = flat_batch_idx * n
+      tmem_slice = x_tmem.slice(slice(0, m), slice(col_start, col_start + n))
+      tmem_slice.store(val_slice)
+  else:
+    x_tmem.store(value)
   return ()
 
 
@@ -3775,11 +3792,30 @@ def _async_store_tmem_lowering_rule_wg(
       handle_transposes=False,
       handle_reshapes=False,
   )
+  batch_shape = ()
+  if transforms and isinstance(
+      transforms[0], gpu_core.ExpandLeadingBatchDimensionsTransform
+  ):
+    batch_shape = transforms[0].batch_shape
+    transforms = transforms[1:]
   if transforms:
     raise NotImplementedError(
         f"Unimplemented transforms for TMEM refs. {transforms=}"
     )
-  mgpu.dialect.async_store_tmem(value, x_tmem)
+  if batch_shape:
+    m, n = ir.VectorType(value.type).shape[-2:]
+    for batch_idx in np.ndindex(batch_shape):
+      flat_batch_idx = int(np.ravel_multi_index(batch_idx, batch_shape))
+      val_slice = vector_dialect.extract(
+          value, dynamic_position=[], static_position=batch_idx
+      )
+      col_start = flat_batch_idx * n
+      tmem_slice = mgpu_utils.memref_slice(
+          x_tmem, (slice(0, m), slice(col_start, col_start + n))
+      )
+      mgpu.dialect.async_store_tmem(val_slice, tmem_slice)
+  else:
+    mgpu.dialect.async_store_tmem(value, x_tmem)
   return ()